moves tx construction to exchange tab
