+++ src/components/atomic-swap/exchange.tsx
@@ -1,5 +1,7 @@
 import React, { ChangeEvent } from "react";
+import BigNumber from "bignumber.js";
 import {
+  BASE_FEE,
   Memo,
   MemoType,
   Operation,
@@ -27,10 +29,16 @@ import {
 import { bc, ChannelMessageType } from "helpers/channel";
 import { copyContent } from "helpers/dom";
 import { NetworkDetails, signTx } from "helpers/network";
-import { getServer, submitTx, assembleTransaction } from "helpers/soroban";
+import {
+  getServer,
+  submitTx,
+  assembleTransaction,
+  getTxBuilder,
+  buildSwap,
+} from "helpers/soroban";
 import { ERRORS } from "../../helpers/error";
 
-type StepCount = 1 | 2 | 3 | 4;
+type StepCount = 1 | 2 | 3 | 4 | 5 | 6;
 
 interface ExchangeProps {
   networkDetails: NetworkDetails;
@@ -42,24 +50,28 @@ interface ExchangeProps {
 export const Exchange = (props: ExchangeProps) => {
   const [contractID, setContractID] = React.useState("");
   const [signedTx, setSignedTx] = React.useState("");
-  const [footprint, setFootprint] = React.useState("");
   const [txResultXDR, setTxResultXDR] = React.useState("");
   const [isSubmitting, setIsSubmitting] = React.useState(false);
   const [stepCount, setStepCount] = React.useState(1 as StepCount);
 
   const [exchangeKey, setExchangeKey] = React.useState("");
 
+  const [tokenAAddress, setTokenAAddress] = React.useState("");
+  const [amountA, setAmountA] = React.useState("");
+  const [minAmountA, setMinAmountA] = React.useState("");
+  const [tokenBAddress, setTokenBAddress] = React.useState("");
+  const [amountB, setAmountB] = React.useState("");
+  const [minAmountB, setMinAmountB] = React.useState("");
+  const [swapperBAddress, setSwapperBAddress] = React.useState("");
+  const [originalFootprint, setOriginalFootprint] = React.useState("");
+
   bc.onmessage = (messageEvent) => {
     const { data, type } = messageEvent.data;
 
     switch (type) {
-      case ChannelMessageType.Footprint: {
-        setFootprint(data.footprint);
-        return;
-      }
       case ChannelMessageType.SignedTx: {
         setSignedTx(data.signedTx);
-        setStepCount(3);
+        setStepCount(5);
         return;
       }
       default:
@@ -100,7 +112,7 @@ export const Exchange = (props: ExchangeProps) => {
 
   function renderStep(step: StepCount) {
     switch (step) {
-      case 4: {
+      case 6: {
         return (
           <>
             <Heading as="h1" size="sm" addlClassName="title">
@@ -132,7 +144,7 @@ export const Exchange = (props: ExchangeProps) => {
           </>
         );
       }
-      case 3: {
+      case 5: {
         const submit = async () => {
           const server = getServer(props.networkDetails);
 
@@ -148,7 +160,9 @@ export const Exchange = (props: ExchangeProps) => {
             tx,
             props.networkDetails.networkPassphrase,
             txSim,
-            xdr.LedgerFootprint.fromXDR(Buffer.from(footprint, "base64")),
+            xdr.LedgerFootprint.fromXDR(
+              Buffer.from(originalFootprint, "base64"),
+            ),
           );
 
           const _signedXdr = await signTx(
@@ -201,11 +215,61 @@ export const Exchange = (props: ExchangeProps) => {
           </>
         );
       }
-      case 2: {
-        const handleChange = (event: ChangeEvent<HTMLInputElement>) => {
-          setContractID(event.target.value);
+      case 4: {
+        const handleSwapperBAddress = (
+          event: ChangeEvent<HTMLInputElement>,
+        ) => {
+          setSwapperBAddress(event.target.value);
         };
-        const goToSwapperA = () => {
+        const handleTokenBChange = (event: ChangeEvent<HTMLInputElement>) => {
+          setTokenBAddress(event.target.value);
+        };
+        const handleTokenBAmountChange = (
+          event: ChangeEvent<HTMLInputElement>,
+        ) => {
+          setAmountB(event.target.value);
+        };
+        const handleTokenBMinAmountChange = (
+          event: ChangeEvent<HTMLInputElement>,
+        ) => {
+          setMinAmountB(event.target.value);
+        };
+
+        const goToSwapperA = async () => {
+          const server = getServer(props.networkDetails);
+          // Gets a transaction builder and use it to add a "swap" operation and build the corresponding XDR
+          const txBuilder = await getTxBuilder(
+            exchangeKey,
+            BASE_FEE,
+            server,
+            props.networkDetails.networkPassphrase,
+          );
+
+          const tokenA = {
+            id: tokenAAddress,
+            amount: new BigNumber(amountA).toNumber(),
+            minAmount: new BigNumber(minAmountA).toNumber(),
+          };
+
+          const tokenB = {
+            id: tokenBAddress,
+            amount: new BigNumber(amountB).toNumber(),
+            minAmount: new BigNumber(minAmountB).toNumber(),
+          };
+
+          const { preparedTransaction, footprint } = await buildSwap(
+            contractID,
+            tokenA,
+            tokenB,
+            exchangeKey,
+            swapperBAddress,
+            "", // memo can be set after rebuild on exchange submit
+            server,
+            props.networkDetails.networkPassphrase,
+            txBuilder,
+          );
+          setOriginalFootprint(footprint);
+
           const newWindow = window.open(
             `${window.location.origin}/swapper-a`,
             "_blank",
@@ -214,10 +278,118 @@ export const Exchange = (props: ExchangeProps) => {
             newWindow.onload = () => {
               bc.postMessage({
                 type: ChannelMessageType.ContractID,
-                data: contractID,
+                data: {
+                  baseTx: preparedTransaction.toEnvelope().toXDR("base64"),
+                  contractID,
+                },
               });
             };
           }
+        };
+
+        return (
+          <>
+            <Heading as="h1" size="sm">
+              Choose Token B
+            </Heading>
+            <Input
+              fieldSize="md"
+              id="swapper-b-address"
+              label="Swapper B Address"
+              value={swapperBAddress}
+              onChange={handleSwapperBAddress}
+            />
+            <Input
+              fieldSize="md"
+              id="token-b-id"
+              label="Token ID"
+              value={tokenBAddress}
+              onChange={handleTokenBChange}
+            />
+            <Input
+              fieldSize="md"
+              id="token-b-amount"
+              label="Amount"
+              value={amountB}
+              onChange={handleTokenBAmountChange}
+            />
+            <Input
+              fieldSize="md"
+              id="token-b-min-amount"
+              label="Min Amount"
+              value={minAmountB}
+              onChange={handleTokenBMinAmountChange}
+            />
+            <div className="submit-row">
+              <Button
+                size="md"
+                variant="tertiary"
+                isFullWidth
+                onClick={goToSwapperA}
+              >
+                Build Swap
+              </Button>
+            </div>
+          </>
+        );
+      }
+      case 3: {
+        const handleTokenAChange = (event: ChangeEvent<HTMLInputElement>) => {
+          setTokenAAddress(event.target.value);
+        };
+        const handleTokenAAmountChange = (
+          event: ChangeEvent<HTMLInputElement>,
+        ) => {
+          setAmountA(event.target.value);
+        };
+        const handleTokenAMinAmountChange = (
+          event: ChangeEvent<HTMLInputElement>,
+        ) => {
+          setMinAmountA(event.target.value);
+        };
+
+        return (
+          <>
+            <Heading as="h1" size="sm">
+              Choose Token A
+            </Heading>
+            <Input
+              fieldSize="md"
+              id="token-a-id"
+              label="Token ID"
+              value={tokenAAddress}
+              onChange={handleTokenAChange}
+            />
+            <Input
+              fieldSize="md"
+              id="token-a-amount"
+              label="Amount"
+              value={amountA}
+              onChange={handleTokenAAmountChange}
+            />
+            <Input
+              fieldSize="md"
+              id="token-a-min-amount"
+              label="Min Amount"
+              value={minAmountA}
+              onChange={handleTokenAMinAmountChange}
+            />
+            <div className="submit-row">
+              <Button
+                size="md"
+                variant="tertiary"
+                isFullWidth
+                onClick={() => setStepCount((stepCount + 1) as StepCount)}
+              >
+                Next
+              </Button>
+            </div>
+          </>
+        );
+      }
+      case 2: {
+        const handleChange = (event: ChangeEvent<HTMLInputElement>) => {
+          setContractID(event.target.value);
         };
         return (
           <>
@@ -236,7 +408,7 @@ export const Exchange = (props: ExchangeProps) => {
                 size="md"
                 variant="tertiary"
                 isFullWidth
-                onClick={goToSwapperA}
+                onClick={() => setStepCount((stepCount + 1) as StepCount)}
               >
                 Go to Swapper A
               </Button>
